requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
pandas==2.2.3
openpyxl==3.1.5
gunicorn==22.0.0
Werkzeug==3.1.3
psycopg2-binary==2.9.9
Flask-Dance==7.1.0
requests-oauthlib==1.3.1
